requirements.txt
pinecone-client
langchain
openai
azure-identity
tiktoken
fastapi
uvicorn